Add refactor the code..
