+++ readme.md
@@ -1,24 +1,43 @@
 中国股票行情数据分析
-=================================================
-写了好久的的Api和各大网站查找才得到一份比较全面的股票数据
-=================================================
+==============
+中国金融数据API
+==============
 
+同时支持Python2, Python3.
 
-### 1.能够获得股票部分金融数据 ###
-### 2.能够获得日线，周线，60分钟数据线，月线，30分钟数据。###
-### 3.能够获得历史数据，但是部分各大网站的股票的历史数据有很大的误差，情谨慎鉴别 ###
-### 4.该项目依赖于httplib2，和beautifulsoup。###
-### 5.python开发环境是2.7.5，其他版本环境没有测试。 ###
-### 6.腾讯大盘行情板块数据有大bug废弃处理
+相对于`Tushare`，希望在对于数据质量的把控上更有优势。对于API的接口的稳定性更可靠。代码能够简单明了。
+接口的数据API都是相对靠谱，如果数据有太大的迁就偏差，就是不能保证数据API的质量。
+互联网金融数据本身质量就有数据存活偏差，对于一些平台大的平台的数据质量，计算的结果完全有很大的偏差。
 
-如果有任何问题，请联系QQ：365504029
+# todo
+   美国证券股市
+   香港证券股市
+   期权行情API
 
+1. 能够获取基础数据。
+
+       - base  证券基础数据
+       - financial 证券财报数据
+       - history 证券历史数据
+       - market 证券市场数据
+       - realtime 证券实时数据
+
+1. 支持`证券基础数据`，`证券财报数据`，`证券历史数据` - (1month,1week,1day,60min,30min,5min,1min) 。
+   支持`指数行情数据`, `证券实时行情`，`事件驱动API`.
+   
+2. 如何用API数据构建历史回测交易体系。
+   
+   
 2016-11-08 更新计划
 
 互联网数据有很多问题，金融的数据质量会不能保证。   
 近期会对api进行新的改动，对数据进行验证和新的接口API请求。  
+
+
+
+如果有任何问题，请联系QQ：365504029
  
-本人会画精力和时间找到尽可能准确的行情和数据接口。  
+
