+++ stockqtl/chinastock/base.py
@@ -1,0 +1,29 @@
+# encoding:utf-8
+
+
+def get_stock_list():
+    """
+    :return:
+    """
+
+def get_all_stock():
+    """
+    得到全部上市公司数据
+    :return:
+    """
+    pass
+
+def get_all_market():
+    """
+    得到所有指数数据
+    :return:
+    """
+    pass
+
+def get_stock_industry():
+    """
+    得到板块数据
+    :return:
+    """
+    pass
+
