+++ stockqtl/chinastock/history.py
@@ -1,0 +1,2 @@
+# encoding:utf-8
+
