+++ stockqtl/common.py
