+++ stockqtl/httpGet.py
